13392: work with token in authContextProvider
