+++ src/components/SessionsPageComponents/hooks/useRole.ts
@@ -1,7 +1,7 @@
-import { useAuthProvider } from '@/providers/AuthProvider/useAuthProvider';
+import { getRoleFromToken } from '@/contexts/AuthContext/AuthContext';
 
 export const useRole = () => {
-  const roles = useAuthProvider().getRoleFromToken();
+  const roles = getRoleFromToken();
 
   return (
     roles?.find((role: string) => role === 'STUDENT' || role === 'TEACHER') ||
