+++ src/contexts/AuthContext/AuthContext.tsx
@@ -40,3 +40,23 @@ export const useAuthContext = () => {
 
   return context;
 };
+
+export const getRoleFromToken = () => {
+  const token = localStorage.getItem('accessToken');
+
+  if (!token) {
+    console.error('Токен не найден в localStorage.');
+    return null;
+  }
+
+  try {
+    const base64Payload = token.split('.')[1];
+    const decodedPayload = atob(base64Payload);
+    const roles = JSON.parse(decodedPayload).roles;
+
+    return roles.length > 0 ? roles : null;
+  } catch (error) {
+    console.error('Ошибка при декодировании токена:', error);
+    return null;
+  }
+};
