+++ src/providers/AuthProvider/AuthProvider.ts
@@ -4,7 +4,6 @@ import {
   LoginResponseData,
   UpdateSessionResponseData,
 } from '@/DTO/AuthDTO.ts';
-import { RoleTypes } from '@/types/Role';
 
 export class AuthProvider {
   instance: AxiosClient;
@@ -23,25 +22,5 @@ export class AuthProvider {
     return this.instance.post<UpdateSessionResponseData>('/auth/token', {
       refreshToken,
     });
-  }
-
-  getRoleFromToken(): RoleTypes[] | null {
-    const token = localStorage.getItem('accessToken');
-
-    if (!token) {
-      console.error('Токен не найден в localStorage.');
-      return null;
-    }
-
-    try {
-      const base64Payload = token.split('.')[1];
-      const decodedPayload = atob(base64Payload);
-      const roles = JSON.parse(decodedPayload).roles;
-
-      return roles.length > 0 ? roles : null;
-    } catch (error) {
-      console.error('Ошибка при декодировании токена:', error);
-      return null;
-    }
   }
 }
